Add more stuff

+++ install.d/01-network-manager.sh
@@ -2,7 +2,11 @@
 set -e
 SCRIPT_PATH="$(dirname "$(readlink -f "$BASH_SOURCE")")"
 SCRIPT_NAME="$(basename "$BASH_SOURCE")"
+source $SCRIPT_PATH/../util.sh
 
-sudo apt -y install network-manager
-sudo systemctl start NetworkManager.service
-sudo systemctl enable NetworkManager.service
+if dpkg -s network-manager &>/dev/null; then log_info "network-manager is already installed"
+else
+  sudo apt -y install network-manager
+  sudo systemctl start NetworkManager.service
+  sudo systemctl enable NetworkManager.service
+fi
